Package organizations
1. Created Command object
2. Revised Task object
3. Organized both object definitions under package
